Move `NoteDataSource` and `NoteRepository` to each package
